+++ crates/teloxide-core/src/adaptors/cache_me.rs
@@ -210,6 +210,7 @@ where
         set_sticker_emoji_list,
         set_sticker_keywords,
         set_sticker_mask_position,
+        get_available_gifts,
         send_invoice,
         create_invoice_link,
         answer_shipping_query,