+++ crates/teloxide-core/src/adaptors/throttle/requester_impl.rs
@@ -197,6 +197,7 @@ where
         set_sticker_emoji_list,
         set_sticker_keywords,
         set_sticker_mask_position,
+        get_available_gifts,
         answer_shipping_query,
         create_invoice_link,
         answer_pre_checkout_query,